DataFile.id is now long

+++ app/om/src/main/java/uk/ac/ebi/fg/annotare2/dao/DataFileDao.java
@@ -25,7 +25,7 @@ import java.util.List;
  */
 public interface DataFileDao {
 
-    DataFile get(int id);
+    DataFile get(long id);
 
     DataFile create(String fileName);
